api: allow flow segment paging using opaque key
Allow the next page link to include a "page" query parameter that allows
the next page to be retrieved. If the TAMS implementations uses "page"
then it will likely leave the "timerange" query parameter unchanged.

Marking this change as a feature rather than a breaking change because
clients should be treating the query parameters of a page link as opaque.

sem-ver: feature